[NBS-44] Implement jam sail adcs system
* [NBS-44_implement_JamSail_AdcsSystem] Adding EKF which comuptes kalman gain

* [NBS-44_implement_JamSail_AdcsSystem] Integrating EKF into JamSail

- Tried to use malkolov manetic field sensor but was not getting
  accuracte results
- Reduced to 7 order EKF but kept comments of previous implementation
- Still need to clearn up but putting this commit here as a reference

* [NBS-44_implement_JamSail_AdcsSystem] Updating EKF to remove malkolov model and adding keplarian propogation

* [NBS-44_implement_JamSail_AdcsSystem] Fixing EKF

* [NBS-44_implement_JamSail_AdcsSystem] Fixing EKF to work reasonabily when their is no noise

* [NBS-44_implement_JamSail_AdcsSystem] Adding features to save test runs

* [NBS-44_implement_JamSail_AdcsSystem] Adding multiple changes

* [NBS-44_implement_JamSail_AdcsSystem] Adding sun sensor vector to EKF

* [NBS-44_implement_JamSail_AdcsSystem] Updating cutoff frequencies

* [NBS-44_implement_JamSail_AdcsSystem] Updating parameters of EKF

* [NBS-44_implement_JamSail_AdcsSystem] adding control algorithms

* [NBS-44_implement_JamSail_AdcsSystem] Adding control algorithms

* [NBS-44_implement_JamSail_AdcsSystem] Adding control algorithms

* [NBS-44_implement_JamSail_AdcsSystem] Adding start up state

* [NBS-44_implement_JamSail_AdcsSystem] Got working

* [NBS-44_implement_JamSail_AdcsSystem] Removing comments

* [NBS-44_implement_JamSail_AdcsSystem] cleaning up

* [NBS-44_implement_JamSail_AdcsSystem] Slight drift with angular velocities and torques

* [NBS-44_implement_JamSail_AdcsSystem] Got working

* [NBS-44_implement_JamSail_AdcsSystem] Fixing typo

* [NBS-44_implement_JamSail_AdcsSystem] fixed issue with quaternion multiplication and added magentorquer

* [NBS-44_implement_JamSail_AdcsSystem] Project Submission Status

+++ SourceCode/GenericLibraries/GLegPoly/PublicFunctions/GLegPoly_legendreArrayPolynomials.c
@@ -8,6 +8,8 @@
  *
  */
 
+#include <stdint.h>
+
 /* Function Includes */
 /* None */
 
@@ -20,16 +22,15 @@
 /* Generic Libraries */
 #include "GConst/GConst.h"
 
-int GLegPoly_legendreArrayPolynomials(
-    double *p_outputArray_out,
-    double  inputValue_in,
-    int     degreeN_in)
+int GLegPoly_legendreArrayPolynomials(double *p_outputArray_out,
+                                      double  inputValue_in,
+                                      uint8_t degreeN_in)
 {
   /* Declare local variables */
-  double currentValue;
-  double previousValue1;
-  double previousValue2;
-  int    n;
+  double  currentValue;
+  double  previousValue1;
+  double  previousValue2;
+  uint8_t n;
 
   for (n = 0; n <= degreeN_in; n++)
   {
