+++ SourceCode/GenericLibraries/GLegPoly/PublicFunctions/GLegPoly_legendrePolynomial.c
@@ -8,6 +8,8 @@
  *
  */
 
+#include <stdint.h>
+
 /* Function Includes */
 /* None */
 
@@ -20,16 +22,15 @@
 /* Generic Libraries */
 #include "GConst/GConst.h"
 
-int GLegPoly_legendrePolynomial(
-    double *p_outputValue_out,
-    double  inputValue_in,
-    int     degreeN_in)
+int GLegPoly_legendrePolynomial(double *p_outputValue_out,
+                                double  inputValue_in,
+                                uint8_t degreeN_in)
 {
   /* Declare local variables */
-  double currentValue;
-  double previousValue1;
-  double previousValue2;
-  int    n;
+  double  currentValue;
+  double  previousValue1;
+  double  previousValue2;
+  uint8_t n;
 
   if (degreeN_in == 0)
   {
